src/requirements.txt
openai
python-dotenv
attrs